disable temporary delete for softdeletes

+++ src/Metable.php
@@ -25,7 +25,9 @@ trait Metable
     {
         // delete all attached meta on deletion
         static::deleted(function (Model $model) {
-            $model->purgeMeta();
+            if (!method_exists(__CLASS__, 'trashed')) {
+                $model->purgeMeta();
+            }
         });
     }
